[day_01] Refactor into smaller units
Both in terms of functionality being refactored into helper functions,
and files being refactored into multiple smaller units.

These are all changes intended to improve maintainability as the
repository grows, and new puzzles and solutions keep getting added.

In this particular case, the changes should make it easier to maintain
puzzles that are split into multiple parts.

Also note that the code was grouped under a new namespace specific to
the given puzzle.