[day_01] Add 'Puzzle'
A helper class for setting up the daily puzzles.

Parses arguments passed to the main executable and determines:
- From which text file to load sample input data (Test or sample)
- From which text file to load answers (Test or sample)

Stores the name of the exercise.
Stores the correct path to the sample data based on the given day.

Allows for parsing and fetching answers from text files.

Answers can be filed, checked against the solution, and printed.