[lib] Bit_grid: Add "Dummy map" and add_map() return objects
Added a new constructor that creates an 'empty' grid object.

A "Dummy map" is created by this constructor, which is basically a
temporary Dyn_bitset added to the internal collection of bitsets.

It has the correct size and can be copied to create the first
actually named map, at which point the dummy gets deleted.

Various helper methods were created to create and query the existence
of the dummy map.

Also, other methods such as get_only_map() were adjusted to check
for a dummy map in addition to exisisting logic.

The `add_map` methods now return a reference to the Dyn_bitset
inserted, and parts of the implementation that use these methods now
use that returned reference for any futher operations.