[day_15] Refactor boths parts
(Note that the "old" version is still present at this point)

Refactored the solution to use a more object oriented approach.
Mostly cleaner, easier to extend, less tailored for the specific
task at hand but any performance loss won't be very noticable.

It's better, but not great:
- The Box objects are tightly coupled to the Warehouse (Grid).
- There are some pretty ugly const_cast's.
- Unnecessary updates of the linked tiles of Box objects.
... Etc.

But good enough for now, additional cleanup to be done next.